Rename datafile/scraper/parser modules
