+++ src/python/ensembl/production/datafile/scraper/parsers/embl.py
@@ -16,7 +16,7 @@ import re
 from typing import Optional
 
 from ..utils import get_group
-from .file_parser import BaseOptMetadata, FileParser, FILE_COMPRESSIONS
+from .base_parser import BaseOptMetadata, FileParser, _FILE_COMPRESSIONS_MAP
 
 
 @dataclass
@@ -44,7 +44,7 @@ class EMBLFileParser(FileParser):
         match = self.FILENAMES_RE.match(metadata["file_name"])
         matched_content_type = get_group("content_type", match)
         match = self.FILE_EXT_RE.match(metadata["file_name"])
-        matched_compression = FILE_COMPRESSIONS.get(get_group("compression", match))
+        matched_compression = _FILE_COMPRESSIONS_MAP.get(get_group("compression", match))
         file_extension = get_group("file_extension", match)
         matched_sorting = get_group("sorted", match)
         compression = (
